Implemented the new version of BOSS Job Submitter

+++ setup/Functions.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -
-# echo "Loading \"${BASH_SOURCE[0]/$(dirname ${BOSS_StarterPack})\/}\""
+# echo "Loading \"${BASH_SOURCE[0]/$(dirname ${BOSS_StarterKit})\/}\""
 
 # * ============================================================================= * #
 # *   DESCRIPTION: Shell script that contains definitions of functions that are   * #
@@ -11,7 +11,7 @@
 # *                source Functions.sh                                            * #
 # * ============================================================================= * #
 
-
+source "${BOSS_StarterKit}/setup/FunctionsPrint.sh"
 
 # * ================================== * #
 # * ------- SCRIPT INITIALISER ------- * #
@@ -123,7 +123,7 @@
 # * ================================ * #
 
 
-	function GetBossPackage()
+	function ImportBossPackage()
 	{
 		if [[ -z "${BesArea}" ]]; then
 			PrintError "Shell variable \$BesArea has not been defined"
@@ -133,7 +133,7 @@
 			PrintError "Shell variable \$BOSSINSTALL has not been defined"
 			return 1
 		fi
-		local currentPath="${pwd}"
+		local currentPath="$(pwd)"
 		cd "${BesArea}"
 		local packageName=""
 		printf "Which package do you want to load from the \$BesArea? "
@@ -144,7 +144,13 @@
 			return 1
 		fi
 		if [[ -d "${BOSSINSTALL}/workarea/${packageName}" ]]; then
-			AskForInput "WARNING: Package \"${packageName}\" already exists in your workarea. REMOVE and OVERWRITE?"
+			PrintWarning "Package \"${packageName}\" already exists in your workarea"
+			AskForInput "REMOVE and OVERWRITE?"
+			if [[ $? != 0 ]]; then
+				echo "--> Import aborted"
+				cd "${currentPath}"
+				return 1
+			fi
 			rm -rf "${BOSSINSTALL}/workarea/${packageName}"
 		fi
 		mkdir -p "${BOSSINSTALL}/workarea/${packageName}"
@@ -157,6 +163,41 @@
 		echo "to .gitignore"
 		cd "${currentPath}"
 	}
+	export ImportBossPackage
+
+
+	function SetOutputFolder()
+	{
+		local outputPath="${1:-./output}"
+		local currentPath="$(pwd)"
+		cd "${BOSS_StarterKit}"
+		local currentTarget="$(readlink "output")"
+		if [[ "${currentTarget}" == "" ]]; then
+			currentTarget=output
+		fi
+		if [[ "${outputPath}" == "${currentTarget}" ]]; then
+			PrintSuccess "Path \"${outputPath}\" is already the output path"
+			return 0
+		fi
+		AskForInput "Move all output from \"${currentTarget}\" to \"${outputPath}\"?"
+		if [[ $? != 0 ]]; then
+			cd "${currentPath}"
+			return 1
+		fi
+		mkdir -p "${outputPath}"
+		mv "${currentTarget}/"* "${outputPath}/" > /dev/null
+		rmdir "${currentTarget}"
+		{ rm -f output; } &> /dev/null
+		{ rmdir output; } &> /dev/null
+		ln -s "${outputPath}" output
+		PrintSuccess "Moved all output from"
+		PrintSuccess "  \"${currentTarget}\""
+		PrintSuccess "to"
+		PrintSuccess "  \"${outputPath}\""
+		PrintSuccess "and set symbolic link:"
+		PrintSuccess "  \"output\" --> \"${outputPath}\""
+	}
+	export SetOutputFolder
 
 
 
@@ -170,6 +211,7 @@
 		local path="${1}"
 		if [ ! -d "${path}" ]; then
 			PrintError "Folder \"${path}\" does not exist [${FUNCNAME[0]}]"
+			return 1
 		fi
 	}
 	export CheckDirectory
@@ -202,6 +244,7 @@
 			numFiles=$(find ${directory} -type f | wc -l)
 			if [ $numFiles != 0 ]; then
 				AskForInput "Remove ${numFiles} files in \"${directory}\"?"
+				[[ $? != 0 ]] && return 1
 			fi
 			rm -rf "${directory}"/*
 		else
@@ -238,6 +281,7 @@
 		local folderToMake="${1}"
 		if [ ! -d "${folderToMake}" ]; then
 			AskForInput "Create folder the following folder?\n\"${folderToMake}\""
+			[[ $? != 0 ]] && return 1
 			mkdir -p "${folderToMake}"
 		fi
 	}
@@ -245,32 +289,30 @@
 
 
 	function CheckIfFolderExists()
-	# Check if a folder exists. Exit the script if it doesn't.
 	{
 		local folderToCheck="${1}"
 		if [ ! -d "${folderToCheck}" ]; then
 			PrintError "Folder \"${folderToCheck}\" does not exist [${FUNCNAME[0]}]"
-			exit
+			return 1
 		fi
 	}
 	export CheckIfFolderExists
 
 
 	function CheckIfBaseDirExists()
-	# Check if the path to a folder exists. Exit the script if it doesn't.
 	{
 		CheckIfFolderExists "$(dirname "${1}")"
+		[[ $? != 0 ]] && return 1
 	}
 	export CheckIfBaseDirExists
 
 
 	function CheckIfFileExists()
-	# Check if a file exists. Exit the script if it doesn't.
 	{
 		local fileToCheck="${1}"
 		if [ ! -s "${fileToCheck}" ]; then
 			PrintError "File \"${fileToCheck}\" does not exist"
-			exit
+			return 1
 		fi
 	}
 	export CheckIfFileExists
@@ -280,33 +322,6 @@
 # * ====================================== * #
 # * ------- TEXT FILE MODIFICATION ------- * #
 # * ====================================== * #
-
-	function DeleteAllEmptyLines()
-	# Delete all empty lines in a file, including files that only contain whitespace characters
-	{
-		# * Import function arguments
-		local fileName=${1}
-		# * Main function: delete all empty lines of the file
-		sed -i '/^\s*$/d' ${fileName} # delete all empty lines of the file
-	}
-	export DeleteAllEmptyLines
-
-
-	function FormatTextFileToCppVectorArguments()
-	# Feed this function a text file, and it will prepend a `\t"` and append a `",` to each line. The comma is ommited for the last line.
-	{
-		# * Import function arguments
-		local fileName=${1}
-		# * Execute function if lines 
-		local numCorrect=$(grep "^\s*\"[^\"]*\",\s*$" "${fileName}" | wc -l) # does not include last line
-		local numLines=$(cat "${fileName}" | sed '/^\s*$/d' | wc -l)
-		if [ $numLines != $(($numCorrect+1)) ]; then
-			DeleteAllEmptyLines ${fileName}
-			sed -i -e "s/.*/\t\"&\",/" ${fileName} # convert lines to C++ vector arguments
-			sed -i "$ s/.$//"          ${fileName} # remove last comma
-		fi
-	}
-	export FormatTextFileToCppVectorArguments
 
 
 	function ChangeLineEndingsFromWindowsToUnix()
@@ -320,139 +335,6 @@
 	export ChangeLineEndingsFromWindowsToUnix
 
 
-	function SplitTextFile()
-	# Feed this function a path to a text file ($1) and it will split up this file into separate files each with a maximum number of lines ($2 -- default value is 10).
-	{
-		# * Import function arguments
-			local fileToSplit="${1}"
-			local maxNLines=10 # default value
-			if [ $# -ge 2 ]; then
-				maxNLines=${2}
-			fi
-		# * Check arguments
-			CheckIfFileExists "${fileToSplit}"
-		# * Extract path, filename, and extension for prefix * #
-			local path=$(dirname "${fileToSplit}")
-			local filename=$(basename "${fileToSplit}")
-			local extension="${filename/*.}"
-			local filename="${filename%.*}"
-			local prefix="${path}/${filename}_"
-		# * Split input file * #
-		rm -f "${prefix}"???".${extension}" #! remove existing files
-		split -d -a3 -l${maxNLines} "${fileToSplit}" "${prefix}"
-		# * Append extension again (if original file has one) * #
-		# shopt -s extglob # for regex
-		for file in $(ls ${prefix}???); do #! number of ? should match the -a3 argument above
-			DeleteAllEmptyLines "${file}"
-			if [ "${filename}" != "${extension}" ]; then
-				mv -f "${file}" "${file}.${extension}"
-			fi
-		done
-		# ! REMOVE ORIGINAL FILE ! #
-			rm "${fileToSplit}"
-	}
-	export SplitTextFile
-
-
-
-# * ====================================== * #
-# * ------- LS INVENTORY FUNCTIONS ------- * #
-# * ====================================== * #
-
-
-	function CreateFilenameInventoryFromDirectory()
-	# Feed this function a path to a directory ($1) and it will list all files within that directory including their absolute paths. This list will be written to text files ($2) with a maximum number of paths per file ($3 -- default is 0, namely no max). If you wish, you can only list files of a certain extension ($4).
-	{
-		# * Import function arguments
-			local inputDirectory="${1}"
-			local outputFile="${inputDirectory//\//_}"
-				if [ "${outputFile:0:1}" == "_" ]; then
-					outputFile="${outputFile:1}"
-				fi
-				outputFile="filenames/${outputFile}.txt"
-			if [ $# -ge 2 ]; then outputFile="${2}"; fi
-			local maxNLines=8 # default value
-			if [ $# -ge 3 ]; then maxNLines=$3; fi
-			local extension="*" # does not look for extensions by default
-			if [ $# -ge 4 ]; then extension="${4}"; fi
-		# * Check arguments
-			CheckIfFolderExists "${inputDirectory}"
-			CreateBaseDir "${outputFile}"
-		# * Get absolute path of the input directory so that `find` lists absolute paths as well
-			cd "${inputDirectory}"
-			inputDirectory="$(pwd)"
-			cd - > /dev/null
-		# * Make an inventory and write to file
-			find "${inputDirectory}" -iname "*.${extension}" > "${outputFile}"
-			DeleteAllEmptyLines "${outputFile}"
-		# * Split the output file if required
-			if [ $maxNLines -gt 0 ]; then
-				SplitTextFile "${outputFile}" ${maxNLines}
-			fi
-	}
-	export CreateFilenameInventoryFromDirectory
-
-
-	function CreateFilenameInventoryFromFile()
-	# Feed this function a path ($1) to a file containing directories and/or file names and it will list all files within those directories including their absolute paths.  This list will be written to text files ($2) with a maximum number of paths per file ($3 -- default is 0, namely no max). If you wish, you can only list files of a certain extension ($4).
-	{
-		# * Import function arguments * #
-			local inputFile="${1}"
-			local outputFile="${2}"
-			local maxNLines=0 # default value
-			if [ $# -ge 3 ]; then maxNLines=${3}; fi
-			local extension="*" # does not look for extensions by default
-			if [ $# -ge 4 ]; then extension=${4}; fi
-		# * Check arguments * #
-			CheckIfFileExists   "${inputFile}"
-			DeleteAllEmptyLines "${inputFile}"
-			CreateBaseDir "${outputFile}"
-			local startPath="$(pwd)"
-		# * Get absolute path of input file * #
-			CdToBaseDir "${inputFile}"
-			inputFile="$(pwd)/$(basename "${inputFile}")"
-			cd - > /dev/null
-		# * Get absolute path of output file * #
-			CdToBaseDir "${outputFile}"
-			outputFile="$(pwd)/$(basename "${outputFile}")"
-			cd - > /dev/null
-		# * Make an inventory and write to file * #
-			CdToBaseDir "${inputFile}" # in case of relative paths
-			echo > "${outputFile}" # empty output file
-			{ cat "${inputFile}"; echo; } | while read line; do
-				# * Check if empty line *
-				if [ "${line}" != "" ]; then
-					# * If line is a file, just add it to the output file
-					if [ -f "${line}" ]; then
-						echo "Added file \"$(basename ${line}\")"
-						echo "${line}" >> "${outputFile}"
-					# * Otherwise, presume line is a directory and add its contents
-					elif [ -d "${line}" ]; then # check if directory exists
-						CreateFilenameInventoryFromDirectory "${line}" "temp.txt" 0 "${extension}"
-						cat "temp.txt" >> "${outputFile}"
-						echo "Added directory \"$(basename ${line}\") ($(cat temp.txt | wc -l) files)"
-						rm "temp.txt"
-					# * Error moessage if nothing
-					else
-						PrintError "WARNING: \"${line}\" does not exist"
-					fi
-				fi
-			done
-		# * Finalise output file * #
-			DeleteAllEmptyLines "${outputFile}"
-			PrintSuccess "\nRead $(cat "${outputFile}" | wc -l) files and directories from file \"$(basename "${inputFile}")\""
-			PrintSuccess "--> output written to \"$(basename "${outputFile}")\"\n"
-		# * Split the output file if required
-			if [ $maxNLines -gt 0 ]; then
-				echo "Splitting text file \"$(basename ${outputFile})\" to max $maxNLines lines each"
-				SplitTextFile "${outputFile}" ${maxNLines}
-			fi
-		# * Go back to starting directory * #
-			cd "${startPath}"
-	}
-	export CreateFilenameInventoryFromFile
-
-
 
 # * =========================== * #
 # * ------- UNIT TESTER ------- * #
@@ -462,7 +344,7 @@
 	function IniTest()
 	{
 		TestName="${1}"
-		StarterPackObjects="${BOSS_StarterPack}/workarea/Analysis/StarterPack/StarterPack-00-00-00"
+		StarterPackObjects="${BOSS_StarterKit}/workarea/Analysis/StarterPack/StarterPack-00-00-00"
 
 		if [[ ${#} != 1 ]]; then
 			PrintError "IniTest function needs 1 argument"
@@ -501,7 +383,7 @@
 	function gitsync()
 	{
 		# * Go to BOSS Afterburner main dir
-		local mainDir="${BOSS_StarterPack}"
+		local mainDir="${BOSS_StarterKit}"
 		cd ${mainDir}
 		if [ $? != 0 ]; then
 			PrintError "Folder \"${mainDir}\" does not exist\" [${FUNCNAME[0]}]"
@@ -616,11 +498,11 @@
 		# * CHECK PARAMETERS * #
 			if [ ! -d "${doxygenDir}" ]; then
 				echo "\e[91mThis repository does not contain a Doxygen directory (\"${doxygenDir}\")\e[0m"
-				exit
+				return 1
 			fi
 			if [ ! -f "${doxygenPath}" ]; then
 				echo "\e[91mThis repository does not contain a Doxygen file \"${doxygenPath}\"\e[0m"
-				exit
+				return 1
 			fi
 
 		# * PREPARE OUTPUT FOLDERS * #
